feat(hicache): implement HostKVPool slot-based host pool (RFC-1.0)
Stage 0 of HiCache: pinned-host KV slot pool providing slot ID
allocation/free and jax.Array reference storage for HiCache L2 tier.

- Move existing PD disaggregation pool to queue_host_kv_pool.py
- Implement RFC-1.0 HostKVPool in host_kv_pool.py (dict + free_list)
- Add unit tests covering all 6 acceptance criteria
